convert tile pos from f32 to i32

+++ src/settings.rs
@@ -3,14 +3,14 @@ use crate::story_time::{
 };
 use bevy::prelude::*;
 
-pub const WW: usize = 1200;
-pub const WH: usize = 800;
+pub const WW: i32 = 1200;
+pub const WH: i32 = 800;
 
-pub const GRID_COLS: f32 = 250.0;
-pub const GRID_ROWS: f32 = 250.0;
+pub const GRID_COLS: i32 = 250;
+pub const GRID_ROWS: i32 = 250;
 
-pub const GRID_COLS_HALF: f32 = GRID_COLS / 2.0;
-pub const GRID_ROWS_HALF: f32 = GRID_ROWS / 2.0;
+pub const GRID_COLS_HALF: i32 = (GRID_COLS as f32 / 2.0) as i32;
+pub const GRID_ROWS_HALF: i32 = (GRID_ROWS as f32 / 2.0) as i32;
 
 pub const TILE_SIZE: f32 = 32.;
